Wind acts a lot more like wind

+++ globals.cpp
@@ -7,6 +7,6 @@ unsigned int gFullScreenOption = 0;
 
 unsigned int gTerrainType = 0;
 unsigned int gObstacleTotal = 0;
-unsigned int gWind = 0;
+double gWind = 0.0;
 unsigned int gFullScreen = 0;
 unsigned int gAI_Enable = 0;
